+++ app.js
@@ -38,87 +38,10 @@ app.set('port', process.env.PORT || 3000)
 
 // ////////////////
 // 路由設定
-app.get('/', (req, res) => {
-  res.redirect('/index')
-})
 
-app.get('/index', (req, res) => {
-  Restaurant.find( (err, restaurants) => {
-    if (err) return console.error(err)
-
-    const js = { delBtn: "delBtn", catch: 'catchError' }
-    res.render('index', { css: 'index', js, restaurants })
-  })
-})
-
-app.get('/restaurants/new', (req, res) => {
-  res.render('newEdit', { new: 'new' })
-})
-
-app.get('/restaurants/:id/edit', (req, res) => {
-  Restaurant.findById(req.params.id, (err, restaurant) => {
-    if (err) return console.error(err)
-    res.render('newEdit', { restaurant })
-  })
-})
-
-app.get('/restaurants/:id', (req, res) => {
-  Restaurant.findById(req.params.id, (err, restaurant) => {
-    if (err) return console.error(err)
-
-    const js = { delBtn: "delBtn", catch: 'catchError' }
-    res.render('show', { css: 'show', js, restaurant })
-  }) 
-})
-
-app.delete('/restaurants/:id/delete', (req, res) => {
-  Restaurant.findById(req.params.id, (err, restaurant) => {
-    if (err) return console.error(err)
-    restaurant.remove(err => {
-      if (err) return console.error(err)
-      res.redirect('/')
-    })
-  })
-})
-
-app.get('/search', (req, res) => {
-  // 將 search keyword 轉成正規表達式來比對餐廳名稱
-  const keyword = req.query.keyword
-  const regexp = new RegExp(keyword, 'i')
-
-  Restaurant.find( (err, collection) => {
-    if (err) return console.error(err)
-
-    // 過濾 name & category
-    const restaurants = collection.filter(
-      item => item.name.match(regexp) || item.category.match(regexp)
-    )
-
-    const js = { delBtn: "delBtn", catch: 'catchError' }
-    res.render('index', { css: 'index', js, restaurants, keyword })
-  })
-})
-
-app.post('/restaurants/new', (req, res) => {
-  const input = req.body
-
-  Restaurant.create(input)
-  res.redirect('/')
-})
-
-app.put('/restaurants/:id/edit', (req, res) => {
-  Restaurant.findById(req.params.id, (err, restaurant) => {
-    if (err) return console.error(err)
-
-    for (const key in req.body) {
-      restaurant[key] = req.body[key]
-    }
-    restaurant.save(err => {
-      if (err) return console.error(err)
-      res.redirect('/restaurants/' + req.params.id)
-    })
-  })
-})
+app.use('/', require('./routers/index.js'))
+app.use('/search', require('./routers/search.js'))
+app.use('/restaurants', require('./routers/restaurants.js'))
 
 // ////////////////
 // 啟動 Server
